requirements.txt
google-generativeai>=0.5.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.7.0
typer>=0.9.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pytz>=2024.1
SQLAlchemy>=2.0.0
httpx>=0.25.0
lxml>=5.0.0
